SPLIT : util 함수 모듈로 구성
